requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
configparser==3.5.0
dnspython==1.15.0
enum34==1.1.6
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.1.0
futures==3.2.0
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
greenlet==0.4.15
httplib2==0.12.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
oauthlib==2.1.0
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==1.9.3
python-engineio==3.0.0
python-socketio==2.1.0
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.14
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
